+++ hydromt_delft3dfm/__init__.py
@@ -2,8 +2,7 @@
 
 from os.path import abspath, dirname, join
 
-__version__ = "0.1.3.dev"
-
+__version__ = "0.2.1.dev"
 DATADIR = join(dirname(abspath(__file__)), "data")
 
 from .dflowfm import *