+++ docs/changelog.rst
@@ -18,6 +18,21 @@ Changed
 Fixed
 -----
 
+v0.2.0 (20 November 2023)
+=========================
+Major dependency upgrade and add support for compound structures.
+
+Added
+-----
+- Support reading and writing of compound structures. (PR#113)
+
+Changed
+-------
+- Upgraded meshkernel dependency to version 3.0.0. (PR#109)
+- Upgraded xugrid dependency to version 0.7.1. (PR#109)
+- Upgraded hydrolib-core dependency to version 0.6.0. (PR#109)
+- Support multiple structures at the same location. (PR#113)
+
 v0.1.2 (20 October 2023)
 ========================
 Added examples, updated documentation and Upgraded dependency.
@@ -28,16 +43,16 @@ Added
 
 Changed
 -------
-- Upgraded hydromt dependency to version 0.9.0. (PR#100) 
+- Upgraded hydromt dependency to version 0.9.0. (PR#100)
 - Updated documentation. (PR #97)
 
 v0.1.1 (13 October 2023)
 ========================
-Dependencies upgrade. 
+Dependencies upgrade.
 
 Changed
 -------
-- Upgraded meshkernel dependency to version 2.1.0. (PR#94) 
+- Upgraded meshkernel dependency to version 2.1.0. (PR#94)
 
 v0.1.0 (22 September 2023)
 ==========================